add helpful error message when passing Intention to dataframe helpers; refactor
