+++ grama/__init__.py
@@ -6,6 +6,7 @@ from .core import *
 # Import dfply
 # --------------------------------------------------
 from .dfply import *
+from .dataframe import *
 
 # Marginals uses make_symbolic decorator
 from .marginals import *
